Automatic SGP30 baseline calibration sequence
If the persisted baseline is is not valid or older than 1 week, the baseline will not be restored to the SGP30. This will result in the sensor trying to determine a new baseline.
The adjustement algorithm will be accelerated for 12hrs which is the Maximum time required to find a new baseline. If the device has been operated for 12h, the baseline will be persisted.